# openapi.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: Envelope Budgeting API
  description: This simple personal finance budgeting API allows users to organize expenses into "envelopes" using the "envelope budgeting" strategy. More background information on this API is available on the project's README at [https://github.com/e-hamilton](https://github.com/e-hamilton).
...
servers:
  - url: https://envelopes-api.herokuapp.com/
    description: Heroku
...
paths:
...
  /expenses:
    get:
      summary: Get all expenses in the system.
      description: Returns a paginated collection of envelopes belonging to all users. Currently, the maximum number of items per page is set to 5 for testing purposes.
      tags:
      - Expenses
      parameters:
      - name: x-access-token
        in: header
        description: Valid JWT for *any* user obtained from POST /auth.
        required: true
        schema:
          type: string
      responses:
        200:
          $ref: "#/components/responses/Success_ExpensesCollection"
        401:
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create a new expense.
      description: Create a new expense belonging to the user associated with the access token in the request.
      tags:
      - Expenses
      parameters:
      - name: x-access-token
        in: header
        description: Valid JWT for *any* user obtained from POST /auth.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - name
              - description
              - cost
              properties:
                name:
                  type: string
                  description: The name of the expense.
                description:
                  type: string
                  description: A description of the expense.
                cost:
                  type: number
                  description: The expense's cost.
            example:
              name: New spacesuit helmet
              description: To stay safe while sitting in a tin can far above the world...
              cost: 250
      responses:
        201:
          description: Expense successfully created.
          headers:
            location:
              description: A direct link to the newly-created resource.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The datastore ID of the newly created resource.
                example:
                  id: "222233332223333"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
...
  /envelopes/{envelopeId}/expenses:
    parameters: 
      - name: envelopeId
        in: path
        description: ID of the envelope to access.
        required: true
        schema:
          type: string
      - name: x-access-token
        in: header
        description: Valid JWT for *any* user obtained from POST /auth.
        required: true
        schema:
          type: string
    get:
      summary: Get an envelope's expenses.
      description: Returns a collection of expenses belonging to the envelope specified in the path.
      tags:
      - Associated Envelopes & Expenses
      responses:
        200:
          $ref: "#/components/responses/Success_ExpensesCollection"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
...
    Success_ExpensesCollection:
      description: Successfully retrieved collection of expenses.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Expenses_Collection'
          example: 
            items:
              - id: "222233332223333"
                name: New spacesuit helmet
                self: https://app.com/expenses/222233332223333
                owner:
                  id: "1234123412341234"
                  self: https://app.com/users/1234123412341234
                envelope:
                  id: "5678567856785678"
                  self: https://app.com/envelopes/5678567856785678
                description: To stay safe while sitting in a tin can far above the world...
                cost: 250
              - id: "3333222233332222"
                name: Protein pills
                self: https://app.com/expenses/3333222233332222
                owner:
                  id: "1234123412341234"
                  self: https://app.com/users/1234123412341234
                envelope:
                  id: "5678567856785678"
                  self: https://app.com/envelopes/5678567856785678
                description: The idea astronaut food.
                cost: 11.22
            count: 2
    BadRequest:
      description: The request body did not match the schema for this endpoint.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: 
            error: "{missing field} required"
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: 
            error: "{Resource} ID 5555444433332222 not found."
    Unauthorized:
      description: The request's x-access-token header was missing or contained an invalid token. If a valid, unexpired token was provided, the user associated with it most likely does not have permission to access one or more of the requested resource(s).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples: 
            Missing Token:
              value:
                error: Header "x-access-token" required.
            Bad Token:
              value:
                error: jwt malformed
            Expired Token: 
              value: 
                error: jwt expired
            Unauthorized User:
              value: 
                error: You are not authorized to make changes to this resource.
  schemas:
...
    Envelope_Truncated:
      type: object
      properties:
        id:
          type: string
          description: An envelope's 16-digit datastore ID. This is returned as a string to prevent truncation in the even an ID contains leading zeroes.
        self:
          type: string
          description: The URL to access this envelope directly (i.e. .../envelopes/{id}).
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message.
    Expense:
      type: object
      properties:
        id:
          type: string
          description: The expense's 16-digit datastore ID. This is returned as a string to prevent truncation in the even an ID contains leading zeroes.
        name:
          type: string
          description: The name of the expense.
        self:
          type: string
          description: The URL to access this resource directly (i.e. .../expenses/{id}).
        owner:
          $ref: "#/components/schemas/User_Truncated"
        envelope:
          $ref: "#/components/schemas/Envelope_Truncated"
        cost:
          type: number
          description: The cost of the expense.
        description:
          type: string
          description: A description of the expense.
...
    Expenses_Collection:
      type: object
      required:
      - items
      - count
      properties:
        items:
          type: array
          description: An array of expenses.
          items: 
            $ref: "#/components/schemas/Expense"
        count:
          type: integer
          description: The total count of all expenses matching the query.
        next:
          type: string
          description: A URL to the next page of results (if applicable).
...
    User_Truncated:
      type: object
      properties:
        id:
          type: string
          description: A user's 16-digit datastore ID. This is returned as a string to prevent truncation in the even an ID contains leading zeroes.
        self:
          type: string
          description: The URL to access this user directly (i.e. .../users/{id}).